Fragments are saved when activity changes
In this version,I added double click on back button to quit the app in order to maintain the user in the app when there is a false click on the back button.

I only saved the current fragment. I didn't take into consideration when the user clicks twice on the a button in the Bottom Navigation Menu (it destroys the fragment and creates it)